requirements.txt
streamlit
streamlit-webrtc
opencv-python
numpy
Pillow